+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Administration/HyperBan.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using KaguyaProjectV2.KaguyaBot.Core.Extensions;
+using KaguyaProjectV2.KaguyaBot.Core.Extensions.DiscordExtensions;
 
 namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Administration
 {