+++ KaguyaProjectV2/KaguyaBot/Core/Services/MusicService.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using Discord;
 using Humanizer;
 using KaguyaProjectV2.KaguyaBot.Core.Extensions;
+using KaguyaProjectV2.KaguyaBot.Core.Extensions.DiscordExtensions;
 using KaguyaProjectV2.KaguyaBot.Core.Services.ConsoleLogServices;
 using KaguyaProjectV2.KaguyaBot.DataStorage.JsonStorage;